feat: support pixelRatio
