--- packages/device/.gitignore
@@ -1,13 +1,0 @@
-*.swp
-*.log
-
-.DS_Store
-.idea/
-package-lock.json
-yarn.lock
-coverage/
-node_modules/
-build/
-dist/
-lib/
-_miniapp/
