+++ packages/device/docs-template/README.md
@@ -19,17 +19,20 @@ import { platform, appName, screenWidth, screenHeight, appVersion } from 'univer
 
 ## APIS
 
-### `appName: String`
+### `appName: string`
 app name or browser name
 
-### `appVersion: String`
+### `appVersion: string`
 app version or browser version info
 
-### `platform: String`
+### `platform: string`
 platform: 'Android'、'iOS'、'MacIntel', etc.
 
-### `screenWidth: Number`
+### `screenWidth: number`
 Screen width in px
 
-### `screenHeight: Number`
+### `screenHeight: number`
 Screen height in px
+
+### `devicePixelRatio: number`
+ The ratio of the resolution in physical pixels to the resolution in CSS pixels for the current display device.(Not supported in Weex)
