+++ packages/device/src/index.ts
@@ -1,8 +1,8 @@
 import { isWeb, isWeex, isMiniApp, isWeChatMiniProgram } from 'universal-env';
 import webModule from './web';
 import weexModule from './weex';
-import miniappModule from './miniapp/ali';
-import wechatModule from './miniapp/wechat';
+import miniappModule from './ali-miniapp';
+import wechatModule from './wechat-miniprogram';
 
 function dutyChain(...fns) {
   for (let i = 0; i < fns.length; i++) {
@@ -53,6 +53,7 @@ const appVersion = deviceInfo.appVersion;
 const platform = deviceInfo.platform;
 const screenWidth = deviceInfo.screenWidth;
 const screenHeight = deviceInfo.screenHeight;
+const devicePixelRatio = deviceInfo.devicePixelRatio;
 
 export {
   appName,
@@ -60,4 +61,5 @@ export {
   platform,
   screenWidth,
   screenHeight,
+  devicePixelRatio
 };
