+++ packages/device/tsconfig.json
@@ -2,24 +2,17 @@
   "compilerOptions": {
     "module": "esNext",
     "target": "es2015",
-    "jsx": "preserve",
-    "jsxFactory": "createElement",
-
+    "outDir": "build",
     "moduleResolution": "node",
-    "baseUrl": ".",
-
-    "rootDir": "src",
-    "downlevelIteration": true,
+    "sourceMap": true,
     "alwaysStrict": true,
-    "outDir": "lib",
-    "strict": true,
-    "preserveConstEnums": true,
-    "experimentalDecorators": true,
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "noImplicitAny": false,
-    "noImplicitThis": false
+    "baseUrl": "."
   },
-  "include": ["src/**/*"],
-  "exclude": ["**/__tests__"]
+  "include": [
+    "./src/**/*"
+  ],
+  "exclude": [
+    "demo",
+    "**/__tests__"
+  ]
 }
